move RemoteRepository into src/workspace.rs
