+++ src/workspace.rs
@@ -1,7 +1,8 @@
 use std::collections::BTreeMap;
 use std::io;
-use std::path::{Path, MAIN_SEPARATOR};
+use std::path::{Path, PathBuf, MAIN_SEPARATOR};
 use walkdir::WalkDir;
+use url::Url;
 
 use config::Config;
 use vcs;
@@ -41,7 +42,7 @@ impl Workspace {
   pub fn clone_repository(&self, query: &str) {
     let url = remote::make_remote_url(query).unwrap();
     if let Some(ref root) = self.config.roots.iter().next() {
-      let repo = remote::RemoteRepository::new(url).unwrap();
+      let repo = RemoteRepository::new(url).unwrap();
       repo.clone(&root, None).unwrap();
     }
   }
@@ -141,5 +142,63 @@ impl Repository {
 
   pub fn relative_path(&self) -> String {
     self.path.clone()
+  }
+}
+
+
+struct RemoteRepository {
+  protocol: String,
+  base_url: String,
+  user: String,
+  project: String,
+}
+
+impl RemoteRepository {
+  pub fn new(url: Url) -> Result<RemoteRepository, String> {
+    let protocol = url.scheme().to_owned();
+    let base_url = try!(url.host_str().ok_or("cannot retrieve host information".to_owned()))
+      .to_owned();
+
+    let paths: Vec<_> = try!(url.path_segments().ok_or("failed to split URL".to_owned()))
+      .map(ToOwned::to_owned)
+      .collect();
+    let user = paths[0].clone();
+    let repo = paths[1].trim_right_matches(".git").to_owned();
+
+    Ok(RemoteRepository {
+      protocol: protocol,
+      base_url: base_url,
+      user: user,
+      project: repo,
+    })
+  }
+
+  fn url(&self) -> Url {
+    Url::parse(&format!("{}://{}/{}/{}.git",
+                        self.protocol,
+                        self.base_url,
+                        self.user,
+                        self.project))
+      .unwrap()
+  }
+
+  fn local_path(&self, root: &str) -> PathBuf {
+    let mut dest = PathBuf::from(root);
+    dest.push(&self.base_url);
+    dest.push(&self.user);
+    dest.push(&self.project);
+    dest
+  }
+
+  pub fn clone(&self, root: &str, depth: Option<i32>) -> Result<(), io::Error> {
+    let url = self.url();
+    let dest = self.local_path(root);
+    if dest.exists() {
+      println!("exists: {}", dest.display());
+    } else {
+      println!("clone: {} -> {}", url.as_str(), dest.display());
+      try!(vcs::Git::clone(url, dest.as_path(), depth));
+    }
+    Ok(())
   }
 }
